+++ src/Geopilot.Api/Context.cs
@@ -44,6 +44,7 @@ public class Context : DbContext
                 .Include(d => d.DeliveryMandate)
                 .Include(d => d.Assets)
                 .Include(d => d.DeclaringUser)
+                .Include(d => d.PrecursorDelivery)
                 .AsNoTracking()
                 .ToList();
         }